backend/api/requirements.txt
asgiref==3.7.2
autopep8==2.0.1
certifi==2022.12.7
chardet==3.0.4
click==8.1.3
colorama==0.4.6
distlib==0.3.7
Django==4.2.5
djangorestframework==3.14.0
dnspython==2.4.2
docopt==0.6.2
email-validator==2.0.0.post2
filelock==3.12.4
Flask==1.1.2
Flask-Cors==3.0.8
greenlet==2.0.1
idna==2.6
itsdangerous==2.1.2
Jinja2==2.11.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
pipreqs==0.4.13
platformdirs==3.10.0
pycodestyle>=2.6.0
pytz==2023.3.post1
requests==2.18.4
six==1.16.0
SQLAlchemy==1.4.45
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.22
virtualenv==20.24.5
Werkzeug==2.2.2
yarg==0.1.9
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
pyJWT==2.3.0
dj_database_url==2.1.0
